requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.5
django-bootstrap-icons==0.8.2
django-bootstrap-v5==1.0.11
django-modeltranslation==0.18.8
idna==3.4
psycopg2-binary==2.9.5
requests==2.28.2
soupsieve==2.3.2.post1
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.26.14
